Izkomentirani kod sa predavanja 14.02.2020. radi pomoci izostalom kolegi

+++ pozarko/7.2.1_krug/Program.cs
@@ -6,16 +6,17 @@ namespace _7._2._1_krug
     {
         static void Main(string[] args)
         {
-            float r = float.Parse(Console.ReadLine());
+            int r = int.Parse(Console.ReadLine());
             Console.WriteLine("Opseg: {0}, Povrsina: {1}",Opseg(r),Povrsina(r));
         }
 
-        private static double Povrsina(float r)
+        //Prima int, vraca double(jer je PI tipa double), da nije mozemo joj staviti tip float jer ne ocekujemo cijelobrojni broj.
+        private static double Povrsina(int r)
         {
             return r * r * Math.PI;
         }
 
-        private static double Opseg(float r)
+        private static double Opseg(int r)
         {
             return 2 * r * Math.PI;
         }
